+++ src/components/NextLevelIndication.tsx
@@ -1,11 +1,13 @@
 import React from "react";
 import { FaArrowRight } from "react-icons/fa6";
 
-function NextLevelIndication(){
-    return <div>
-    <FaArrowRight />
-    Next level
-  </div>;
+function NextLevelIndication() {
+  return (
+    <div className="flex gap-1 font-extrabold items-center">
+      <FaArrowRight />
+      Next level
+    </div>
+  );
 }
 
-export default NextLevelIndication
+export default NextLevelIndication;