+++ src/components/PlayAgainButton.tsx
@@ -28,7 +28,7 @@ function PlayAgainButton({ render }: Props) {
     <>
       {!gameInProgress ? (
         <a href="#" onClick={handleClick} className="fixed top-20 left-20">
-          <span className="text-white flex gap-2">
+          <span className="text-white flex gap-2 items-center">
             <RiRestartLine /> Play again
           </span>
         </a>